feat(organizations): add theming/story for Mantine Loader and LoadingOverlay TASK-1464 (#5425)
### 📣 Summary
Theming for Mantine's Loader and LoadingOverlay was added to match
Kobo's current LoadingSpinner for general purpose.

### 📖 Description
Both 'regular' and 'big' variant were added with its respective type
names to be used with Mantine's `Loader` and `LoadingOverlay`
components.
To simplify the implementation I decided to simply use the existing
`LoadingSpinner` component instead of extracting it's contents.
Since Mantine's `Loader` doesn't have the `message` parameter, right now
I'm fixing it on `false` so it won't appear. We may need to implement a
Wrapper in the future to add a message, depending on our future needs.

* `LoadingOverlay` is already in use at
https://github.com/kobotoolbox/kpi/pull/5414


### 👀 Preview steps
- Use storybook to check the visual of the component and compare with
the existing loader in use in the app.