fix(billing): plan page always redirecting to settings TASK-1473 (#5440)
### 📣 Summary
Fixes a bug that was causing the plans page to always redirect current
user to account settings page.

### 💭 Notes
There was a redundant check for org ownership in the page component. It
seems to be triggering even if user *is* an org owner, probably because
of a race condition. But we handle ownership checks elsewhere already,
so I am simply removing the code.

### 👀 Preview steps
As an org owner on a Stripe-enabled instance, navigate to the plans
page. On `main` you will be redirected to account settings. On this
branch the plans page should successfully load.