feat: back-fill submission unique identifier - TASK 1459 (#5434)
### 📣 Summary
Added a long-running migration to backfill the `root_uuid` field for old
submissions.


### 📖 Description
Submissions now use a field called `root_uuid`, which serves as a unique
identifier per project. For older submissions, this field is empty. This
PR introduces a long-running migration to backfill the `root_uuid` field
with the current `uuid` for existing submissions.

To ensure uniqueness, the migration includes logic to detect and resolve
conflicts. If a `uuid` is not unique within a project, the
`clean_duplicated_submissions` management command is invoked to address
the duplicates before assigning the `root_uuid`.