chore: upgrade pyxform TASK-1467 (#5442)
### 💭 Notes
Developer-facing changes only. 
Pxyform 3.0.0 has a number of breaking changes. Most notable for our
purposes are the refactoring of `SurveyElement`s into a bunch of
subclasses and the removal of the `get_abbreviated_xpath` method, which
we made extensive use of.

The refactoring of `SurveyElement` mostly resulted in a lot of Attribute
Not Found errors as, for example, `Survey `objects no longer have a
`bind` property (when everything was a `SurveyElement`, that property
would just be an empty dict). In all places where this broke things, we
updated the code to either check the existence of the attribute first or
check the class of the object.

To deal with the removal of `get_abbreviated_path,` we just recreated
the method within the repo, using the `iter_ancestors` method that
replaced `get_lineage`.

It turns out some of the files touched had a lot of formatting issues
(mostly double quotes instead of single), which the formatter picked up
and fixed, resulting in several changes that are not actually part of
the upgrade. Hopefully those are relatively obvious.