feat(organizations): Add isEmpty to avatar component and use it for members table TASK-1420 (#5457)
### 📣 Summary
Ghost variant was added to display a "ghost" state in members table when
user is invited but still not active

### 📖 Description
- Added a `isEmpty` prop for our existent `Avatar` component for it to
render the empty-dashed circle.
- At first I intended to add theme and use Mantine's Avatar component,
but since our avatar has so many specific details the effort wouldn't be
worth it, so I just customized the existent component.
- In the future we can migrate the internal structure of our avatar
component to use Mantine's layout components.
- Avatar story was changed to add the isEmpty as a config option

### 👀 Preview steps
One can check the implementation result in Avatar's storybook story
For reference, I'm adding a screenshot of the property in use:

![image](https://github.com/user-attachments/assets/70b8b366-8c26-49ac-aff9-cff27cda8ffa)