ci: failed Github Actions notify Zulip (#5484)
### 💭 Notes

For clarity and simpler notification logic, also separated `ci-pr.yml`
and `ci-main.yml` along the way.

### 👀 Preview steps

- see the deleted
[`ci-test.yml`](https://github.com/kobotoolbox/kpi/pull/5484/commits/808bb5794c55d72ef59fc2e0e9e43f41ccc11623)
that's a copy of main's CI file except for branch being this branch
instead of `main`. It should work exactly the same but on the main.
- see the Zulip message to preview the produced message. The message and
links are dynamically set, and so should be correct on the main as well.