feature(organizations): add functionality to the MMO invite modal TASK-1373 (#5493)
### 📣 Summary
<!-- Delete this section if changes are internal only. -->
<!-- One sentence summary for the public changelog, worded for
non-technical seasoned Kobo users. -->

Adds functionality to the invite modal to invite one member at a time to
your MMO.

### 📖 Description
<!-- Delete this section if summary already said everything. -->
<!-- Full description for the public changelog, worded for non-technical
seasoned Kobo users. -->

User can use a username or email for the recipient. Email will work
without the recipient having a Kobo account (it will prompt them to sign
up).

### 👷 Description for instance maintainers
<!-- Delete this section if everything is already said above. -->
<!-- Full description for the public changelog, worded for technical
Kobo instance maintainers. -->

This PR includes a POC for the `useForm` extension to mantine. It is a
great way to use mantine components with built in validation so we don't
need to mess around with juggling a bunch of states. NOTE: currently
only `useField` has asynchronous validation and only works on one
component, not an entire form. The concepts are exactly the same though
so it still serves as a POC.

### 💭 Notes
<!-- Delete this section if empty. -->
<!-- Anything else useful that's not said above,worded for
reviewers, testers, and future git archaeologist collegues. Examples:
- screenshots, copy-pasted logs, etc.
- what was tried but didn't work,
- conscious short-term vs long-term tradeoffs,
- proactively answer likely questions,
-->

Invites will have strange effects on the members table until #5480 is
merged.

The check right now to differentiate usernames from emails is only a
simple `.includes('@')` as per the spec on the notion task. This does
not match the criteria for the backend to not process the request as a
user. This means sending an invite to `something@notanemail` will pass
the validation on the frontend but still fail on the backend and result
in a "Failed to send invite" notify.

A `return await` needed to be added to the queries in
`membersInviteQuery.ts` in order to have proper error handling. See
[chat](https://chat.kobotoolbox.org/#narrow/stream/4-Kobo-Dev/topic/Teams.20.2F.20organizations/near/548349).

### 👀 Preview steps
<!-- Delete this section if behavior can't change. -->
<!-- If behavior changes or merely may change, add a preview of a
minimal happy path. -->

1. Be a user with permissions to invite to their MMO.
2. Enable the feature flag `?ff_orgMemberInvitesEnabled=true`
3. Notice that either a valid username or a string matching an email
structure is required alongside the role selection for the "Send invite"
button to enable
4. Notice a row added to the members table to confirm the invite has
been sent

---------

Co-authored-by: Paulo Amorim <pauloamorimbr@gmail.com>
Co-authored-by: James Kiger <68701146+jamesrkiger@users.noreply.github.com>