fix(organization): ensure data transfer occurs upon invite acceptance TASK-1556 (#5519)
### 📣 Summary
Fixed an issue where data transfer was not triggered when an invite was
accepted.


### 📖 Description
A bug prevented data from being transferred when an invited user
accepted an organization invitation. This fix ensures that upon
acceptance, the transfer process is correctly triggered, allowing data
associated with the user to be reassigned to the organization as
intended.