Regenerate snapshots when downloading XForm XML (#5128)
## Description

Invokes pyxform every time "Download XML" is clicked, which avoids
serving stale XForm XML in certain situations. For example, prior to
this work, if a form had *not* been changed since its XForm XML was last
generated, but the behavior of pyxform *had* changed, clicking "Download
XML" would have ignored the pyxform changes.

Please note that the XForm XML obtained from "Download XML" is *not*
identical to the XForm XML generated when deploying a form. Deploying or
redeploying a form was *not* affected by this stale XML issue, even
prior to this change.