fix(NLP): missing suggested language in language selector DEV-17 (#5559)
### 💭 Notes
Loop of asynchronous calls was bad. I made Promise.all that fixes the
problem.

### 👀 Preview steps
I can reproduce this on kf.main but I believe this is on production too.

1. Create a form with one audio question
2. Create two submissions (each with one audio file)
3. From the data table, click Open for the first submission.
4. Create a manual transcription in English. Save it.
5. Delete the transcript.
6. Go to the second submission (from the data table or by just clicking
the up/down button from the same screen).
7. Create a manual transcription in Spanish. Save it.
8. Go back to the first subscription. Try to create a new transcript in
English.

Expected behavior: The top of the language list shows English in the
first position.

Instead: English is not displayed at all (and can't be found by
searching either). Spanish is displayed in first position.

---------

Co-authored-by: Akuukis <Akuukis@users.noreply.github.com>