fix(assets): do not update date_modified when viewing asset page TASK-1672 (#5581)
### 📣 Summary
Fixes a bug where the most recent modify date on assets was changing
when the page was viewed.


### 💭 Notes
Updates the save_to_db method to take an additional parameter to include
`date_modified` in the fields to update. The field defaults to True
since most of the time it is called when we are actually setting a value
from a user action and do want the date_modified to update. When we are
just getting the enketo id, though, sets it to False.


### 👀 Preview steps

Bug template:
1. ℹ️ have an account and a project
2. In the Django shell, run
`my_asset.deployment.save_to_db({'enketo_id':None})` and
`my_asset.refresh_from_db`
3. Look at `my_asset.date_modified` and record the value
4. Navigate to the project page for the asset
5. In the Django shell, refresh the asset and look at the date_modified
6. 🔴 [on main] The date_modified will have been updated
7. 🟢 [on PR] The date_modified will not have been updated