feat(attachments): create query file for removing attachments and a way to test it TASK-570 TASK-572 (#5577)
### 💭 Notes
This has multiple different changes in it, so that it is much easier to
test and review :)

The changes:
- Added upcoming endpoints to `api.endpoints.ts`
- Defined `removingAttachmentsEnabled` feature flag 🏴 
- Created `AttachmentActionsDropdown` component
- This component is meant to be used in all places that display an
attachment. It has an option to "Download" or "Delete" attachment.
- It has a stories file that mostly works - notifications don't appear,
and attachment is not being actually deleted, but you can get the
general vibe of the component, and test `is_deleted` flag with it (see
`submissionData` arg description)
- Note: it is being hidden behind feature flag, so you need to set it
for Storybook too
- It has a mocks file that have an asset object with single submission
with single attachment 👌
- Uses `AttachmentActionsDropdown` inside Single Submission Modal, so
that it can be tested in real scenario (apart from testing in Storybook)
- Improved `SubmissionResponse` interface to make it clearer which uuid
we should use
- Also changed `deployment__data_download_links`, as it turns out not
all `ExportTypeName`s can happen in there :)

In the next PRs I am going to:
1. Use `AttachmentActionsDropdown` in the rest of the places (here I
only add it in one place)
2. Write code for handling attachments deletion in Submission Modal
(here I only left a comment)

### 👀 Preview steps

Testing:
1. ℹ️ Have account and a project with `image` question and a submission
2. Go to Project -> Data -> Table and open Single Submission Modal (eye
button)
- Note: enable `removingAttachmentsEnabled` feature flag (feature flag
needs to be present in the url when you're opening the modal, because
hook checks it when it runs during component start)
3. 🟢 Notice that "…" button appears next to the image
4. Click "…"
5. 🟢 Notice that "Download" and "Delete" options are present
- Note: if your user doesn't have `change_submissions` permission,
"Delete" will be disabled
6. 🟢 Upon clicking "Download", file will be downloaded
7. Click "Delete" option
8. 🟢 Notice a modal appearing with "Cancel" and "Delete" buttons
9. Click "Delete" button (it uses a mocked call)
- 33% times an error will happen (error notification, and modal stays
open)
- 66% times a success will happen (success notification, and modal
closes)

---------

Co-authored-by: Anji Tong <ollejna@gmail.com>