feat(attachments): add download and delete buttons TASK-571 (#5599)
### 💭 Notes

We don't touch the Project > Data > Gallery in MVP, so this would be the
only place that we don't give option to delete attachments from.

Changes here:
- `AttachmentActionsDropdown` now displays "Delete" button in red color
(Tessa's request)
- `AttachmentActionsDropdown` uses `<a>` for "Download" button for
better UX
- `AudioPlayer` has option to render things in "right section"
- `SidebarSubmissionMedia` uses `AttachmentActionsDropdown`
- `SidebarSubmissionMedia` doesn't handle "video" question types anymore
(we weren't using this option)
- `MediaCell` (and `TableMediaPreview`) now handles "file" question
types
- `MediaCell` now displays filename next to the icon
- Reasoning: previously `file` was displaying filename there, and it is
definitely useful to see it. I decided to change this for the other
types as well, as it is better to see some different text next to icon
rather than the identical icon for all submissions
- `MediaCell` uses `AttachmentActionsDropdown`
- Created `markAttachmentAsDeleted` utility function that adds
`is_deleted` flag to given attachment in a submission data object
- `SubmissionDataTable` now reacts to attachment being deleted, and
causes two things:
1. memoized submission data in `SubmissionModal` is being updated (using
`markAttachmentAsDeleted`)
2. `DataTable` is force refreshing submissions, so that it renders the
deleted attachment (this will be visible once `is_deleted` flag is being
handled)
- Fixed `DataTable` listener for
`actions.resources.refreshTableSubmissions`, as we broke it recently
(probably I did it when migrating to TS)
- Fixed `TableBulkCheckbox` default export (it worked previously, but
now Biome was complaining)

### 👀 Preview steps

Case A: testing steps for Submission Modal:
1. ℹ️ Have a project with `audio`, `file`, `image`, and `video` question
2. Make a submission that has responses to all the questions
- go to Project > Form and use "Open" button in the "Collect data"
section
3. Go to Project > Data > Table
4. Open submission ("eye" icon button)
5. 🟢 Notice that each question response has now "…" button on the right
side
- Using "Download" button (link) should work for each of question type
   - Using "Delete" button should do this for each question type:
1. A 33% chance of action failing (mock) would cause error notification
to appear
2. A 67% chance of action suceeding (also mock) would cause success
notification to appear, "…" modal to be closed, and memoized attachment
data in Submission Modal getting `is_deleted` flag
6. 🟢 Notice that when attachment `is_deleted` the "…" button is not
rendered
- Note that the attachment would still appear, as the work on displaying
`DeletedAttachment` component is still in progress

Case B: testing steps for Table Media Preview modal (`file`, `image`,
and `video` questions):
1. Go through steps 1-3 of Case A
2. 🟢 Notice in the table that cell now has a file name and an icon
button (previously it was just the file name for `file` and just icon
for the other two)
3. Open a modal (icon button in the cell)
4. 🟢 Notice in the modal that we now have "…" button in top right
instead of "Download" button
   - Using "Download" button (link) should work
   - Using "Delete" button should do this for each question type:
1. A 33% chance of action failing (mock) would cause error notification
to appear
2. A 67% chance of action suceeding (also mock) would cause success
notification to appear, "…" modal to be closed, and Data Table
submissions to be refetched
5. 🟢 Notice that for `file` type the file name and mimetype are being
displayed as modal content

Case C: testing steps for NLP View:
1. Go through steps 1-3 of Case A
2. For `audio` question response use "Open" button
4. 🟢 Notice that in the right side the audio player now has "…" button
   - Using "Download" button (link) should work
   - Using "Delete" button should do this:
1. A 33% chance of action failing (mock) would cause error notification
to appear
2. A 67% chance of action suceeding (also mock) would cause success
notification to appear, "…" modal to be closed, and submission data to
be refetched

---------

Co-authored-by: Anji Tong <ollejna@gmail.com>