feat(massEmail): update mass email models TASK-679 (#5608)
### 📣 Summary
Added support for scheduling mass emails with a configurable frequency
and introduced a constraint to prevent duplicate email records for the
same user in a given email job.



### 📖 Description
- Added `frequency` (integer) and `live` (boolean) fields to
`MassEmailConfig` to allow configuring email sending behavior.
- Ensured `frequency` defaults to `-1` and `live` defaults to `False`.
- These fields are not displayed in the admin panel.
- Enforced a uniqueness constraint on `(email_job, user)` in
`MassEmailRecord` to prevent duplicate email records for the same user
in a given email job.