fix: add activity log implementation into release branch 2.025.02 TASK-1714 (#5618)
### 📖 Description
We need to add the implementation code for the activity logs feature
existing in main into the release branch 2.025.02.
Front end and back end code will be added via this PR
The release branch was created at a point in time where only half of the
implementation was done and now we have a bug in the release branch
because of the missing half and inconsistency between front and backend
codes.